jwt done

+++ controllers/imageController.js
@@ -1,7 +1,8 @@
 import Image from "../models/imageModel.js";
+import User from "../models/userModel.js";
 
 export const getImages = async (req, res) => {
-  const images = await Image.find();
+  const images = await Image.find({ userId: req.user.id });
   return res.status(200).json(images);
 };
 
@@ -13,6 +14,7 @@ export const createImage = async (req, res) => {
 
   const image = await Image.create({
     title: req.body?.title,
+    userId: req.user.id,
   });
   res.status(201).json(image);
 };
@@ -22,6 +24,18 @@ export const updateImage = async (req, res) => {
   if (!image) {
     res.status(400);
     throw new Error("Provide id");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (image.userId.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
   }
 
   const updatedImage = await Image.findByIdAndUpdate(req.params.id, req.body, {
@@ -35,6 +49,17 @@ export const deleteImage = async (req, res) => {
   if (!image) {
     res.status(400);
     throw new Error("Provide id");
+  }
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (image.userId.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User Not authorized");
   }
 
   await image.deleteOne();
